Implement expiration of Bearer token after 24 hours
Was a breeze to implement. But a nightmare to test, because of
auto_now_add on the created field of the Token.
No warning from Django when trying to change the field.
Just silently ignoring my changes in the database. This sucks.
Took me at least two hours, where implementing the feature lasted 15
minutes.