Refactoring: use a core app for all non api things
We started with an api app, and put all our models in it.
We now use a core app and an api app. Only api related things are in the
latter now.

We moved the migrations and deleted the database because we don't have
any sensible data yet. Otherwise we should have used this more tedious
way: https://realpython.com/move-django-model/